Remove check for X11
The only thing that was using it was a preprocessor branch in
kkeysequencewidget.cpp, which only had branches for Mac and X11.  The
restriction to those two was because it originally used KKeyServer to
get the modifier names, and that was only implemented for X11 and Mac.
This is no longer relevant.

REVIEW: 115024