Preserve translation domain when merging .rc files
When two .rc documents that specify different translation domains are
merged, only one of the two domains will appear in the merged document.
This breaks translation of text elements under the dropped domain.
To fix it, the top domain attribute (or the application domain) is now
propagated to each text element before merging. Then, at the moment
of translation, the local domain attribute takes priority over
the top attribute (or the application domain).

REVIEW: 122937
BUG: 342976