Fix shortcuts when secondary shortcut set.
If a user defined secondary ("alternate") shortcut is set, an issue in
KXMLGUIFactory causes reloading of both primary and secondary from the
RC file's DOM to fail, resulting in no effective shortcuts available for
the action. Fix by checking for the correct type in the shortcut
property for QActions.

REVIEW: 123508
BUG: 345411