Add dedicated Version tab to KAboutApplicationDialog
This replaces the version information of the version and frameworks and
moves it into a dedicated tab. In this tab the information is extended
by the Qt version (which is equally relevant as e.g. the frameworks
version) in both runtime and compile time.

Also windowing system is added. This will become a useful information
for KWin developers starting in Plasma 5.4 when users start to test
things out and we need to know whether the window they experience the
problem with is running on wayland or xwayland.

REVIEW: 124133
CHANGELOG: About application dialog shows dedicated "Version" tab with further information.