KHelpMenu: mDonateAction is already parented to the KHelpMenu.
Don't delete it manually. This fixes a crash during shutdown of
KMix.
This would be just unnecessary if the action was actually
deleted during destruction of the parent set in the constructor,
which happens later, but according to valgrind, at least in KMix
the action is reparented to and deleted by some action collection
that is cleared earlier. "If you take it, it's yours."