Warn on startup about ambiguous shortcuts (with an exception for Shift+Delete)
Add a warning at the createGui stage about ambiguous shortcuts being found in the same action collection.

This is usually a developer issue, but the error message about ambiguity will only show up when someone tries to use the shortcut, so it is relatively easy to miss if you do not try all your actions via a shortcut.

Also if the involved shortcut is one of the non primary shortcuts of edit_cut + deletefile, just give it away, since it's Shift+Delete being fought over as part of our defaults being ambiguous.

REVIEW: 129299