KKeySequenceWidget: make it possible to record Ctrl+Num+1 as a shortcut.
Summary:
Qt supports this string representation for Qt::KeypadModifier since 5.1
but this code didn't support that modifier.

This is the easy part of the fix for the 8-years-old bug, and it makes
it possible to use the numeric keypad for application shortcuts.

The bigger issue is handling of keypad modifiers in global shortcuts
(patches to come for kwindowsystem and kglobalaccel).

CCBUG: 183458

Test Plan: Set shortcut Ctrl+Num+1 for "decrement number by 1" action in kate, works.

Reviewers: aacid

Reviewed By: aacid

Subscribers: apol, #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D6226