Do not allow to configure separator actions via context menu
Summary:
It's not useful to add shortcuts for separators, nor to add them
to the toolbar. This patch prevents the confusing user interaction.
Note that the menu will instead be hidden, just like when you'd
left-click on a separator. But I guess this strange behavior should
be changed in Qt upstream instead (to keep the menu open).

Reviewers: dfaure

Subscribers: #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D10860