Remove visibilityChanged connection in favor of existing eventFilter
KMainWindow compresses saving its configuration by a 500ms timer. However, for QDockWidget visibilityChanged
a direct connection to the setSettingsDirty method is made, which immediately saves settings.
Since there is already an eventFilter also listening for QEvent::Show and QEvent::Hide, the connection to
visibilityChanged is redundant since that signal is emitted from the respective event handler anyway.
The main benefit is that the eventFilter code compresses the configuration save,
not jeopardizing startup time for applications with lots of dock widgets like Dolphin.

Differential Revision: https://phabricator.kde.org/D22523