fix default shortcut detection
Summary:
When the default shortcut is empty the UI string "None" is used to
indicate that. Alas, when mapping QKeySequences to the UI label we have to
rely on QKS::toString being equal/notequal to the text in the label and
that is not the case since an empty QKS will be an empty string, but
our display string is "None".

This resulted in the "custom" sequence detection incorrectly thinking all
shortcuts without a default are customized when in fact they aren't, the
strings are just different because we express it as "None" and QKS as "".
To mitigate this we'll simply ensure empty sequences are serialized to our
None reference string instead of an empty string.

Test Plan:
broken master:
{F7662310}

with fix:
{F7662311}

also tested continued working of:
- setting a custom key
- unsetting a custom key

Reviewers: dfaure, aacid

Reviewed By: aacid

Subscribers: aacid, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D24908