Port away from KTextWidgets
Summary:
It is only used for the bugreport dialog for non-KDE apps. By replacing KTextEdit with QTextEdit we can get rid of a rather large dependency with minimal impact.

The only complication is that the krichtexteditor test uses it. For more context see T12195.

To work around this we only look for KTextWidgets if actually building the tests and only include the relevant test if ktextwidgets is found

Test Plan: builds

Reviewers: #frameworks, apol

Reviewed By: apol

Subscribers: apol, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D25629