remove ability to report bugs against arbitrary stuff from a static list
Summary:
this list hasn't been updated since 2013.
David Edmundson also told me that removing it was discussed recently, so
here we go! the combobox is simply replaced with a static qlabel for the
current kaboutdata, working under the assumption that when you click
report bug you want to report one for the current application not all
the stuff on bugs.kde.org

Test Plan: it builds and it can report against kwrite

Reviewers: davidedmundson, dfaure

Reviewed By: dfaure

Differential Revision: https://phabricator.kde.org/D26019