[RFC] Save window state outside of XDG_CONFIG_HOME
The window state is volatile data and thus should not be saved in .config

This moves the default save location from .config/appnamerc to .local/share/appname/appnamestaterc, which is more appropriate.

This is particularly relevant when .config is version-controlled where you don't want volatile changes, i.e. when moving a window

CCBUG: 416731