Allow opting out of remembering window positions on X11
Various people have requested this, for reasons that seem sensible
enough to me:
- Some people like the KWin positioning modes and want all windows to
  follow them
- It might be annoying to have only KDE apps follow this setting and
  preferable to not use it at all for pepole who use mostly non-KDE apps
- The per-screen-arrangement memory feature may interact strangely or in
  a buggy manner for some people's screen arrangements

For those reasons, it seems reasonable to allow disabling the feature,
though it still remains on by default. This commit turns it off if
`AllowKDEAppsToRememberWindowPositions=false` is set in the user's
`kdeglobals` file. A GUI to toggle this will be added in a separate
commit to some System Settings KCM.

CCBUG: 415150