[kmainwindow] Don't delete entries from an invalid kconfiggroup
If for some reason the config group we're trying to delete entries from
doesn't exist, the app will assert in KConfigGroup::deleteEntry in
`Q_ASSERT_X(isValid(), "KConfigGroup::deleteEntry", "accessing an
invalid group");`

So let's guard against that by only deleting the entry from the config
group if the group itself is valid. Maybe it's not, for some reason.

BUG: 427236
FIXED-IN: 5.75