Inhibit keyboard shortcuts while recording
On wayland it is not possible to grab the keyboard. This means that it is
currently not possible to reassign key combinations that are currently in use
because the shortcut is triggered instead. Fortunately there exists a wayland
protocol to tell the compositor that we do not want it to trigger shortcuts
while the surface has the keyboard focus.
CCBUG:407395