Make ToolTips expandable
This commit adds a small hint to tooltips which can be expanded.
Pressing Shift will show the expanded help in a widget that isn't
hidden when the mouse moves. This is necessary so we can have
clickable links there which will allow us to refer users to the
manual, community wiki etc. which makes these resources more readily
available and helps us avoid duplicating their work. It also allows
us to keep the tooltips a reasonable length.

If a widget provides an expanded tooltip but no normal tooltip, not
even a hint is shown when hovering the widget. This way no hovering
text will appear unless it was explicitly set by a developer. At
the same time this lets us keep Whats This available for all areas.

An exception to this are QMenus. The hint to show the expanded
tooltip is shown for them even if no tooltip has been explicitly
set. If a tooltip has been explicitly set, that information is also
shown because this implies that the iconText() which is
automatically set to be the toolTip() is not sufficient.

This commit also adds the shortcut to tooltips of `QToolButton`s.

Having tooltips be expandable lets us minimise the risk of tooltips
covering currently used areas of an application by accident.

Related discussion: https://phabricator.kde.org/T9986