Do not show a tooltip for actions that have a submenu
Do not show a tooltip when there is a menu since they will compete
space-wise. It would also be weird if you hover an action and on
the one side a tooltip appears and on the other a submenu. When
both show up users might have trouble figuring out where to look.