Fix tooltips sometimes not showing up in menus
Qt only posts QEvent::ToolTips if the cursor was moved *after* a
QMenu was hidden. This is problematic because when the cursor is
moved above a QMenu then submenus appear and hide all the time.

This commit makes it so QEvent::ToolTips are posted for QMenus
even if the cursor came to a standstill before a submenu hid.