Remove the Attica requirement by using the store's new links
This patch removes the hard requirement on Attica, but retains most of the functionality (minus the location magic and multiple links), by using the KDE Store's new abilities to link to profiles and avatars using the username of a person.

This adds a checkbox (similar to the one found in Kirigami.AboutPage as of https://invent.kde.org/frameworks/kirigami/-/merge_requests/283) which defaults to off, so we never go online to fetch stuff unless the user explicitly allows for this to happen. Future work might be to store this checkbox' state, perhaps in a global fashion (discussion in comments suggests probably not).

This is for the the https://phabricator.kde.org/T12198 task