KShortcutsDialog: add another, simpler, constructor
Working on porting away from deprecated KShortcutsDialog methods, it looks
like the dialog is mostly constructed with the default args (KShortcutsEditor::AllActions
and KShortcutsEditor::LetterShortcutsAllowed), so it makes sense to add a
constructor that's easier to use than having to use that boilerplate everywhere;
and the first constructor is still there for specialised usage where you
don't want to show all action types, or not allow single letter shortcuts.

Also fix a code example, the first arg of the first constructor is not
a KActionCollection, but rather a KShortcutsEditor::ActionTypes.