commandbar: collect actions using the guiFactory
Using `KActionCollection::allCollections` is a really bad idea here
because it can lead to actions with same names from different clients
being present in the KCommandBar and the user has no way to tell which
action is from which client. In Kate, that means for each opened tab you
get tons of actions with same name, leading to a really unusable
experience for some actions.

BUG: 439526
Signed-off-by: Waqar Ahmed <waqar.17a@gmail.com>