Add the menubar actions to the action collection
These are the File, Edit, Settings ...etc top level menu actions on the menu
bar; this is so that those menus can be found by KActionCollection::action(name).

One actual use-case is in Konsole, it has a "Plugins" menu, and if no plugins
were loaded, we'd want to hide that menu from the menubar; right now this
is done by matching the QAction objectName() which is set to the "name="
attribute in the ui.rc XML file.