Update translation context of two strings
This commit slightly improves the i18n call by adding contexts
which were previously erroneously placed within the comment.

This kind of change is normally not a great idea because it forces
all translators to re-visit this string but in this case this is
the wanted result:

There was previously an error in these calls which made the
multiline comment explaining how to translate these strings not
show up which lead to unfortunate translations. With this all fixed
now, translators re-visiting this string as part of the their usual
workflow and without being pinged explicitly e.g. in a mailing list
seems like the least amount of disruption which causes these
translation to be fixed.