Allow KToolBar to be in other places than MainWindow's ToolBarArea
Without this change there was a crash when appling changes made through the "Configure Toolbars" dialog (KEditToolBar) if you had KToolBars in other places than the QMainWindow's Qt::ToolBarArea.

After fixing the crash the toolbar however was always moved from its position to the Qt::TopToolBarArea. This gets fixed by choosing a "clear, hide and show" approach for toolbars that are not direct children of the QMainWindow. For others we leave it to the current "delete and recreate" approach.