Fix testAutoSaveSettings
fc46a2215809abebfdab8ca471c1b27b65b7ad56 in KConfig changed the config
file keys for width and height saving to no longer be surrounded by
spaces on both sides. This commit adjusts the test to compensate, and
loosens the matching condition slightly to hopefully prevent further
changes to the key format from breaking the test--so long as they
continue to contain the text "Width" and "Height", which they will for
the foreseeable future.