Clean up unsafe/unused copytext leftovers
Having code that calls QProcess with a random string from the UI is quite unsafe.
Also, the feature was broken anyway, because the args were not properly split, see 75c29fb8.
Since nobody seemed to care in 3 years, I think it is fine to remove it. That is what we concluded at the KF6 weekly too.

Task: https://phabricator.kde.org/T12208