Add TabFocus to focusPolicy instead of overriding it
Overriding the previous focus policy may reduce the number ways the
widget can receive focus, leading to problems. For example, in
Konqueror, it lead to the user being unable to give focus to the
location bar by simply clicking on it (see
https://mail.kde.org/pipermail/kde-devel/2023-October/002082.html).
To avoid this, `TabFocus` should be added to the existing focus methods.