Add TabFocus to focusPolicy instead of overriding it
This is a new attempt at solving the issue in !203 without causing the compilation error. This is the code I originally wrote in !203, which I later changed according to the suggestion of a reviewer without checking (by mistake) whether the suggested code actually worked.

The reason I used `StrongFocus` instead of just adding `| TabFocus` (aside from the compilation error) is that, unless I'm not understanding the documentation correctly, `ClickFocus|TabFocus` is not a valid value for `Qt::FocusPolicy` which, unlike other similar enum, doesn't have a corresponding `QFlags`.