kmainwindow: Do not save menu bar visibilty when global menu is in use
KXmlGui tracks the visiblity of the QMenuBar widget to try to know if a
user has explicitly shown/hidden a menu.

When a native menu (global menu bar plasmoid) is in use the widget is
hidden, but not because the user explicitly wanted to not show the menu.

If a native menu bar is in use, it does not make sense to have a
save/restore the user preference for the menu.

BUG: 439266