KHelpMenu: Remove unnecessary member variables
createActions() is called by all constructors and calling it a second time
has no effect. Hence, the call in menu() is superfluous. After removing
this call the member variable mActionsCreated is superfluous.

The member variable mShowWhatsThis is superfluous because it's only used
in createActions() and we can pass its value directly to this function.