KAbstractAboutDialog: Do not insert a widget or layout twice
createPersonLayout() lays out the display items according to the
information that is available, using lazy creation to avoid creating
unnecessary widgets or layouts.  Unfortunately, if the person item
being created has both a web address and an email address, the 'name'
or 'col' is inserted twice into the 'row' layout near the end of each
of their conditional blocks.  This results in a display with stretched
blank space around the tool buttons instead of them being right aligned.

The code was originally simplfied (to not use a model and item view) by
commit 31a0756a.

This change simplifies the logic and avoids having to track whether the
'name' or 'col' has already been inserted, by unconditionally creating and
using the 'row' layout in all cases.  Therefore the 'name' or 'col' can be
inserted just once at its intended place.  This does have an overhead of
creating the 'row' when it may not be needed (that is, when the person has
neither a web nor an email address), but this is a rare case and not in a
critical code path so should not be significant.

The same happens, and the equivalent change is implemented, in
KAbstractAboutDialogPrivate::createComponentWidget().