Fix writing default ToolBarsMovable value
Toolbars are locked by default, i.e. ToolBarsMovable is Disabled

If there's no system-wide default and we are set to the default then we
want to call revertToDefault(). Currently it's the other way around.

This doesn't cause functional differences, but results in a redundant
ToolBarsMovable=Disabled being written into the config file