Fix popup transient parents
Currently, the KXMLGUIBuilder forces the QMainWindow as the transient
parent for QMenu objects. The age of this code is more than 20 years.

In Qt 6, a normal QWidget can be a parent of a QMenu. In that case, the
QMenu will walk the widget tree and use the nearest widget with a
native window as the transient parent.

This change relaxes the restrictions in the KXMLGUIBuilder, so the
QMenu can do the right thing for us. Note that a menu can be the
toplevel element. In that case, the container widget is used as the
transient parent.

CCBUG: 500978

Co-authored-by: Christoph Cullmann <christoph@cullmann.dev>