switchlanguage: push country stripped language code to the back of the list
BUG: 509902

If the $LANGUAGE is pt_BR.UTF-8, the QLocale().uiLanguages() returns
QList("pt_Latn_BR", "pt_BR", "pt_Latn", "pt").
We check the translation support of each language code one by one.
If the language is unsupported, we strip the country from the code,
then check the new language code again.

In the pt_BR.UTF-8 case, the first language code "pt_Latn_BR" is
unsupported, after stripping, the code became "pt", which is supported.

This behavior is incorrect, because the result language list is
[pt, pt_BR] instead of [pt_BR, pt].

Push the country stripped language code to the back of the list ensure
the proper precedence.